add creator protect video route and show videos
still no static css changed

+++ controllers/videoController.js
@@ -1,5 +1,6 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import { deserializeUser } from "passport";
 
 export const home = async(req, res) => {
     try{
@@ -10,6 +11,7 @@ export const home = async(req, res) => {
         res.render("home", { pageTitle: "Home", videos: [] });
     }
 };
+
 export const search = async(req, res) => {
     const {query:{ term: searchingBy }} = req;
     let videos=[];
@@ -22,7 +24,9 @@ export const search = async(req, res) => {
     }
     res.render("search", { pageTitle: "Search", searchingBy, videos });
 };
+
 export const getUpload = (req, res) => res.render("upload", { pageTitle: "Upload" });
+
 export const postUpload = async(req, res) => {
     const {
         body:{ title, description },
@@ -31,8 +35,11 @@ export const postUpload = async(req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -41,7 +48,7 @@ export const videoDetail = async(req, res) => {
         params: {id}
     } = req;
     try {
-        const video = await Video.findById(id);
+        const video = await (await Video.findById(id)).populate("creator");
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch(error) {
         console.log(error);
@@ -50,13 +57,18 @@ export const videoDetail = async(req, res) => {
     const video = await Video.findById(id);
     res.render("videoDetail", { pageTitle: "Video Detail" })
 };
+
 export const getEditVideo = async(req, res) => {
     const {
         params: {id}
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video })
+        if(video.creator !== req.user.id){
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch(error) {
         console.log(error);
         res.redirect(routes.home);
@@ -81,7 +93,11 @@ export const deleteVideo = async(req, res) => {
         params: { id }
     } = req;
     try {
-        await Video.findOneAndRemove({ _id: id });
+        if(video.creator !== req.user.id){
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch(error) {
         console.log(error);
     }
